# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dipeat API",
    "description": "찍먹 — 메뉴판 사진 한 장을 구조화된 메뉴로",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/menu/scan": {
      "post": {
        "tags": [
          "menu"
        ],
        "summary": "메뉴판 사진 1장 → 메뉴 목록",
        "description": "목록에 필요한 것만 반환한다. 긴 설명·알레르기 근거는 `/menu/item/explain` 으로.",
        "operationId": "scan_menu_api_v1_menu_scan_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_menu_api_v1_menu_scan_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menu/item/explain": {
      "post": {
        "tags": [
          "menu"
        ],
        "summary": "메뉴 1개의 상세 설명 (카드를 탭했을 때)",
        "description": "사진을 다시 보내지 않는다 — 텍스트 전용이라 목록 호출보다 훨씬 싸고 빠르다.\n\n클라이언트는 결과를 scan_id 와 함께 캐시해서 같은 메뉴를 두 번 묻지 않는다.",
        "operationId": "explain_item_api_v1_menu_item_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/menu/locate": {
      "post": {
        "tags": [
          "menu"
        ],
        "summary": "장바구니 항목을 메뉴판 사진 위에서 찾기 (사진에서 확인 탭)",
        "description": "사용자가 '사진에서 확인' 탭을 열 때만 호출한다. 대상은 장바구니의 몇 개뿐이라\n목록 스캔처럼 항목 수로 곱해지지 않는다. 좌표는 Gemini 네이티브 0~1000 을 받아 0~1 로 변환해 내려준다.",
        "operationId": "locate_items_api_v1_menu_locate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_locate_items_api_v1_menu_locate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_locate_items_api_v1_menu_locate_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "메뉴판 사진 (스캔 때와 같은 축소본)"
          },
          "targets": {
            "type": "string",
            "title": "Targets",
            "description": "찾을 항목 JSON. 예: [{\"index\":1,\"name_local\":\"ラフテー\",\"section\":\"돼지고기 요리\"}]"
          }
        },
        "type": "object",
        "required": [
          "image",
          "targets"
        ],
        "title": "Body_locate_items_api_v1_menu_locate_post"
      },
      "Body_scan_menu_api_v1_menu_scan_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "메뉴판 사진 (JPEG/PNG/WEBP)"
          },
          "mode": {
            "type": "string",
            "enum": [
              "poster",
              "booklet",
              "kiosk"
            ],
            "title": "Mode",
            "description": "촬영 모드",
            "default": "poster"
          },
          "traveler_lang": {
            "type": "string",
            "enum": [
              "ko",
              "ja"
            ],
            "title": "Traveler Lang",
            "description": "여행자 UI 언어",
            "default": "ko"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_scan_menu_api_v1_menu_scan_post"
      },
...
      "ExplainRequest": {
        "properties": {
          "name_local": {
            "type": "string",
            "maxLength": 200,
            "title": "Name Local",
            "description": "1단계 응답의 name_local 을 그대로"
          },
          "name_translated": {
            "type": "string",
            "maxLength": 200,
            "title": "Name Translated",
            "default": ""
          },
          "source_lang": {
            "type": "string",
            "maxLength": 35,
            "title": "Source Lang",
            "description": "1단계 응답의 source_lang",
            "default": "ja"
          },
          "traveler_lang": {
            "type": "string",
            "enum": [
              "ko",
              "ja"
            ],
            "title": "Traveler Lang",
            "description": "여행자 UI 언어",
            "default": "ko"
          },
          "cuisine_hint": {
            "type": "string",
            "maxLength": 200,
            "title": "Cuisine Hint",
            "description": "가게 성격. 있으면 설명이 정확해진다",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name_local"
        ],
        "title": "ExplainRequest"
      },
      "ExplainResponse": {
        "properties": {
          "romanization": {
            "type": "string",
            "title": "Romanization",
            "description": "현지어 발음의 로마자 표기. 모르면 빈 문자열."
          },
          "pronunciation_guide": {
            "type": "string",
            "title": "Pronunciation Guide",
            "description": "여행자 언어 문자로 적은 현지어 발음 안내. 한국어면 한글, 일본어면 가나."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "어떤 음식인지 2~3문장. 맛·식감·조리법·먹는 방법 중심. 호불호가 갈릴 요소(쓴맛, 강한 향신료, 높은 도수, 생식)는 반드시 언급."
          },
          "tip": {
            "type": "string",
            "title": "Tip",
            "description": "여행자에게 도움이 될 조언 한두 문장. 주문 요령, 양, 곁들임 등. 해당 사항이 없으면 빈 문자열."
          },
          "allergens": {
            "items": {
              "$ref": "#/components/schemas/LikelyAllergen"
            },
            "type": "array",
            "title": "Allergens",
            "description": "메뉴명·재료로 정당화할 수 있는 것만. 근거를 댈 수 없으면 넣지 말 것."
          },
          "name_local": {
            "type": "string",
            "title": "Name Local"
          },
          "pronunciation_ko": {
            "type": "string",
            "title": "Pronunciation Ko",
            "description": "구버전 클라이언트 호환 필드. pronunciation_guide 와 같은 값.",
            "deprecated": true
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "romanization",
          "pronunciation_guide",
          "description",
          "tip",
          "allergens",
          "name_local",
          "pronunciation_ko",
          "model",
          "latency_ms"
        ],
        "title": "ExplainResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemBox": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name_local": {
            "type": "string",
            "title": "Name Local"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "x": {
            "type": "number",
            "title": "X",
            "description": "왼쪽. 이미지 너비 대비 0~1 (좌상단 원점, 세운 이미지 기준)"
          },
          "y": {
            "type": "number",
            "title": "Y",
            "description": "위쪽. 이미지 높이 대비 0~1"
          },
          "w": {
            "type": "number",
            "title": "W",
            "description": "너비 0~1"
          },
          "h": {
            "type": "number",
            "title": "H",
            "description": "높이 0~1"
          }
        },
        "type": "object",
        "required": [
          "index",
          "name_local",
          "found",
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "ItemBox",
        "description": "한 항목의 정규화 위치(0~1). 서버가 Gemini 의 0~1000 을 변환해 내려준다."
      },
      "LikelyAllergen": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "egg",
              "milk",
              "buckwheat",
              "peanut",
              "soy",
              "wheat",
              "gluten",
              "mackerel",
              "fish",
              "crab",
              "shrimp",
              "crustacean",
              "squid",
              "shellfish",
              "mollusk",
              "pork",
              "beef",
              "chicken",
              "peach",
              "tomato",
              "sulfite",
              "walnut",
              "pine_nut",
              "tree_nut",
              "sesame",
              "celery",
              "mustard",
              "alcohol",
              "other"
            ],
            "title": "Code",
            "description": "표준 알레르기 코드"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "여행자 언어 표시명"
          },
          "inferred": {
            "type": "boolean",
            "title": "Inferred",
            "description": "메뉴판에 명시돼 있지 않고 요리 지식으로 추론했으면 true. 대부분 true 다."
          },
          "basis": {
            "type": "string",
            "title": "Basis",
            "description": "그렇게 판단한 근거를 여행자 언어 한 문장으로."
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence",
            "description": "확신도"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "inferred",
          "basis",
          "confidence"
        ],
        "title": "LikelyAllergen",
        "description": "⚠️ 메뉴판에서 '읽은' 사실이 아니라 요리 지식에서 '추론한' 값이다.\n\nUI 는 반드시 추정임을 고지해야 한다. 식품 안전 문제라 이 구분이 중요하다."
      },
      "LocateResponse": {
        "properties": {
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/ItemBox"
            },
            "type": "array",
            "title": "Boxes"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "boxes",
          "model",
          "latency_ms"
        ],
        "title": "LocateResponse"
      },
      "MenuItemSummary": {
        "properties": {
          "name_local": {
            "type": "string",
            "title": "Name Local",
            "description": "메뉴판에 적힌 원문 그대로. 절대 번역·정규화하지 말 것. 장음부호(ー)나 가나 표기도 보이는 대로. 사용자가 이 글자를 점원에게 그대로 보여준다."
          },
          "name_translated": {
            "type": "string",
            "title": "Name Translated",
            "description": "요청에서 지정한 여행자 언어의 번역명. 음차가 자연스러우면 음차."
          },
          "price_text": {
            "type": "string",
            "title": "Price Text",
            "description": "가격을 적힌 그대로. 예: '970円', '¥970', '時価'. 못 읽었으면 빈 문자열."
          },
          "price_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Amount",
            "description": "price_text 에서 파싱한 숫자만. 범위·시가·판독 불가면 null. 소수점이 **적혀 있을 때만** 소수로: '$3.50'→3.5, '970円'→970(970.0 이 아니라 970)."
          },
          "tax_included": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Included",
            "description": "税込이면 true, 税抜/+税면 false, 표기가 없으면 null."
          },
          "category": {
            "type": "string",
            "enum": [
              "food",
              "drink",
              "dessert",
              "set",
              "unknown"
            ],
            "title": "Category",
            "description": "분류"
          },
          "section": {
            "type": "string",
            "title": "Section",
            "description": "이 항목이 메뉴판에서 어느 분류 아래 적혀 있는지, 그 분류 제목을 여행자 언어로. 예: '튀김', '찬푸르', '국물', '130엔 접시'. **메뉴판에 실제로 적힌 분류만 쓸 것** — 분류 제목이 없는 메뉴판이면 빈 문자열. 직접 분류를 만들어내지 말 것."
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "이게 무슨 음식인지 한 줄로. **25자 이내.** 예: '흑설탕에 조린 삼겹살'. 긴 설명은 여기 쓰지 말 것 — 사용자가 카드를 탭하면 따로 받아온다."
          },
          "image_query": {
            "type": "string",
            "title": "Image Query",
            "description": "이 요리의 참고 사진을 위키미디어 커먼즈에서 찾기 위한 영문/로마자 검색어 한 개. 가장 널리 통용되는 요리명으로. 예: 'tamagoyaki', 'goya champuru', 'okinawa soba', 'gurukun fried fish'. 재료+조리법 조합이면 그렇게(예: 'fried island scallion'). 커먼즈 파일 제목이 대부분 영문이라 원문보다 영문이 잘 맞는다. 모르면 빈 문자열."
          },
          "tags": {
            "items": {
              "type": "string",
              "enum": [
                "signature",
                "local",
                "caution",
                "spicy",
                "raw",
                "vegetarian",
                "vegan",
                "pork",
                "beef",
                "chicken",
                "seafood",
                "alcohol",
                "share",
                "single_portion",
                "noodle",
                "rice",
                "soup",
                "fried",
                "dessert"
              ]
            },
            "type": "array",
            "title": "Tags",
            "description": "해당하는 것만. 없으면 빈 배열."
          },
          "allergens": {
            "items": {
              "type": "string",
              "enum": [
                "egg",
                "milk",
                "buckwheat",
                "peanut",
                "soy",
                "wheat",
                "gluten",
                "mackerel",
                "fish",
                "crab",
                "shrimp",
                "crustacean",
                "squid",
                "shellfish",
                "mollusk",
                "pork",
                "beef",
                "chicken",
                "peach",
                "tomato",
                "sulfite",
                "walnut",
                "pine_nut",
                "tree_nut",
                "sesame",
                "celery",
                "mustard",
                "alcohol",
                "other"
              ]
            },
            "type": "array",
            "title": "Allergens",
            "description": "메뉴명·재료로 정당화할 수 있는 알레르기 코드만. 근거를 댈 수 없으면 넣지 말 것. 억지로 채우지 말 것. (근거 문장은 상세 조회에서 따로 받는다)"
          },
          "ocr_confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Ocr Confidence",
            "description": "이 항목의 글자를 얼마나 확실히 읽었는지"
          }
        },
        "type": "object",
        "required": [
          "name_local",
          "name_translated",
          "price_text",
          "price_amount",
          "tax_included",
          "category",
          "section",
          "summary",
          "image_query",
          "tags",
          "allergens",
          "ocr_confidence"
        ],
        "title": "MenuItemSummary",
        "description": "결과 목록 카드에 바로 필요한 것만. 여기에 필드를 늘리면 항목 수만큼 곱해진다."
      },
      "MenuScanResponse": {
        "properties": {
          "menu_found": {
            "type": "boolean",
            "title": "Menu Found",
            "description": "이 사진에 **실제로 읽을 수 있는 메뉴판 글자**가 있으면 true. 빈 벽·바닥·하늘·사람·풍경·음식 사진처럼 메뉴판이 아니거나, 아무것도 없는 단색 화면이거나, 너무 어둡거나 흐려서 글자를 하나도 못 읽겠으면 false. **false 면 items 는 반드시 빈 배열이어야 한다.**"
          },
          "no_menu_reason": {
            "type": "string",
            "title": "No Menu Reason",
            "description": "menu_found 가 false 일 때만, 사진에 무엇이 보이는지 여행자 언어 한 문장으로. 예: '아무것도 없는 흰 화면이에요', '너무 어두워 글자가 안 보여요'. menu_found 가 true 면 빈 문자열."
          },
          "source_lang": {
            "type": "string",
            "title": "Source Lang",
            "description": "메뉴판 언어의 BCP-47 코드. 예: 'ja', 'th', 'vi'"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "ISO 4217 통화 코드. 예: 'JPY'. 모르면 빈 문자열."
          },
          "restaurant": {
            "$ref": "#/components/schemas/Restaurant"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MenuItemSummary"
            },
            "type": "array",
            "title": "Items",
            "description": "메뉴판에서 읽은 항목 전부. **사진에서 글자로 읽어낸 것만.** 그럴듯한 메뉴를 지어내는 것은 이 작업에서 가장 심각한 실패다 — 읽을 게 없으면 빈 배열. 카테고리 제목(예: '一品料理')이나 가격대 제목(예: '皿 一三〇円')은 항목이 아니므로 제외. 단, 가격대 제목 아래 나열된 항목들에는 그 가격을 각각 채워 넣을 것."
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "사용자에게 알려야 할 문제를 여행자 언어로. 문제가 없으면 빈 배열."
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "traveler_lang": {
            "type": "string",
            "enum": [
              "ko",
              "ja"
            ],
            "title": "Traveler Lang",
            "default": "ko"
          },
          "meta": {
            "$ref": "#/components/schemas/ScanMeta"
          }
        },
        "type": "object",
        "required": [
          "menu_found",
          "no_menu_reason",
          "source_lang",
          "currency",
          "restaurant",
          "items",
          "warnings",
          "scan_id",
          "meta"
        ],
        "title": "MenuScanResponse"
      },
      "Restaurant": {
        "properties": {
          "name_local": {
            "type": "string",
            "title": "Name Local",
            "description": "사진에서 읽은 가게 이름 원문. 없으면 빈 문자열."
          },
          "name_translated": {
            "type": "string",
            "title": "Name Translated",
            "description": "여행자 언어 표기. 없으면 빈 문자열."
          },
          "cuisine_hint": {
            "type": "string",
            "title": "Cuisine Hint",
            "description": "음식 종류를 여행자 언어로 한 마디. 모르면 빈 문자열."
          }
        },
        "type": "object",
        "required": [
          "name_local",
          "name_translated",
          "cuisine_hint"
        ],
        "title": "Restaurant"
      },
      "ScanMeta": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "image_px": {
            "type": "string",
            "title": "Image Px"
          }
        },
        "type": "object",
        "required": [
          "model",
          "latency_ms",
          "image_px"
        ],
        "title": "ScanMeta"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}